refactor: 빌더 위도, 경도 double 타입으로 수정

+++ src/main/java/com/example/fashionforecastbackend/weather/domain/Weather.java
@@ -106,8 +106,8 @@ public class Weather extends BaseTimeEntity {
 		Season season,
 		SkyStatus skyStatus,
 		RainType rainType,
-		int nx,
-		int ny
+		double nx,
+		double ny
 	) {
 		this.baseDate = baseDate;
 		this.baseTime = baseTime;
